# internal/library/port/web/productweb/product_openapi_types_gen.go
// Package productweb provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package productweb

// Product defines model for Product.
type Product struct {
	CategoryRef *string    `json:"categoryRef,omitempty"`
	Id          *string    `json:"id,omitempty"`
	Name        *string    `json:"name,omitempty"`
	Variants    *[]Variant `json:"variants,omitempty"`
}

// ProductCreated defines model for ProductCreated.
type ProductCreated struct {
	Id *string `json:"id,omitempty"`
}

// Variant defines model for Variant.
type Variant struct {
	Code  *string  `json:"code,omitempty"`
	Id    *string  `json:"id,omitempty"`
	Name  *string  `json:"name,omitempty"`
	Price *float32 `json:"price,omitempty"`
}

// ProductIdParameter defines model for ProductIdParameter.
type ProductIdParameter string

// GetAllProductsSuccessResponse defines model for GetAllProductsSuccessResponse.
type GetAllProductsSuccessResponse []Product

// GetProductByIdSuccessResponse defines model for GetProductByIdSuccessResponse.
type GetProductByIdSuccessResponse Product

// ProductCreationSuccessResponse defines model for ProductCreationSuccessResponse.
type ProductCreationSuccessResponse ProductCreated

// CreateProductJSONBody defines parameters for CreateProduct.
type CreateProductJSONBody Product

// CreateProductJSONRequestBody defines body for CreateProduct for application/json ContentType.
type CreateProductJSONRequestBody CreateProductJSONBody